+++ packages/sign-client/src/controllers/engine.ts
@@ -106,6 +106,9 @@ import {
   getSignDirectHash,
   LimitedSet,
   getWalletSendCallsHashes,
+  getNamespacesChains,
+  getNamespacesMethods,
+  getNamespacesEvents,
 } from "@walletconnect/utils";
 import EventEmmiter from "events";
 import {
@@ -297,12 +300,37 @@ export class Engine extends IEngine {
       id: payloadId(),
       ...((authentication || walletPay) && {
         requests: {
-          authentication: authentication?.map((auth) => ({
-            ...auth,
-            aud: auth.uri,
-            version: "1",
-            iat: new Date().toISOString(),
-          })),
+          authentication: authentication?.map((auth) => {
+            const {
+              domain,
+              chains,
+              nonce,
+              uri,
+              exp,
+              nbf,
+              type,
+              statement,
+              requestId,
+              resources,
+              signatureTypes,
+            } = auth;
+            const protocolParams: AuthTypes.AuthenticateParams = {
+              domain,
+              chains,
+              nonce,
+              type: type ?? "caip122",
+              aud: uri,
+              version: "1",
+              iat: new Date().toISOString(),
+              exp,
+              nbf,
+              statement,
+              requestId,
+              resources,
+              signatureTypes,
+            };
+            return protocolParams;
+          }),
           walletPay,
         },
       }),
@@ -336,6 +364,8 @@ export class Engine extends IEngine {
       }
     });
 
+    await this.setProposal(proposal.id, proposal);
+
     await this.sendProposeSession({
       proposal,
       publishOpts: {
@@ -346,9 +376,11 @@ export class Engine extends IEngine {
           correlationId: proposal.id,
         },
       },
+    }).catch((error) => {
+      this.deleteProposal(proposal.id);
+      throw error;
     });
 
-    await this.setProposal(proposal.id, proposal);
     return { uri, approval };
   };
 
@@ -434,9 +466,6 @@ export class Engine extends IEngine {
       peerPublicKey,
     );
 
-    const { authentication, walletPayResult } =
-      this.prepareProposalRequestsResponses(proposalRequestsResponses);
-
     const sessionSettle = {
       relay: { protocol: relayProtocol ?? "irn" },
       namespaces,
@@ -445,8 +474,7 @@ export class Engine extends IEngine {
       ...(sessionProperties && { sessionProperties }),
       ...(scopedProperties && { scopedProperties }),
       ...(sessionConfig && { sessionConfig }),
-      authentication,
-      walletPayResult,
+      proposalRequestsResponses,
     };
     const transportType = TRANSPORT_TYPES.relay;
     event.addTrace(EVENT_CLIENT_SESSION_TRACES.subscribing_session_topic);
@@ -476,6 +504,8 @@ export class Engine extends IEngine {
       },
       controller: selfPublicKey,
       transportType: TRANSPORT_TYPES.relay,
+      authentication: proposalRequestsResponses?.authentication,
+      walletPayResult: proposalRequestsResponses?.walletPay,
     };
 
     await this.client.session.set(sessionTopic, session);
@@ -499,6 +529,7 @@ export class Engine extends IEngine {
           },
           tvf: {
             correlationId: id,
+            ...this.getTVFApproveParams(session),
           },
         },
       });
@@ -569,7 +600,14 @@ export class Engine extends IEngine {
     }
     const { topic, namespaces } = params;
 
-    const { done: acknowledged, resolve, reject } = createDelayedPromise<void>();
+    const {
+      done: acknowledged,
+      resolve,
+      reject,
+    } = createDelayedPromise<void>(
+      FIVE_MINUTES,
+      "Session update request expired without receiving any acknowledgement",
+    );
     const clientRpcId = payloadId();
     const relayRpcId = getBigIntRpcId().toString() as any;
 
@@ -611,7 +649,14 @@ export class Engine extends IEngine {
 
     const { topic } = params;
     const clientRpcId = payloadId();
-    const { done: acknowledged, resolve, reject } = createDelayedPromise<void>();
+    const {
+      done: acknowledged,
+      resolve,
+      reject,
+    } = createDelayedPromise<void>(
+      FIVE_MINUTES,
+      "Session extend request expired without receiving any acknowledgement",
+    );
     this.events.once(engineEvent("session_extend", clientRpcId), ({ error }: any) => {
       if (error) reject(error);
       else resolve();
@@ -793,7 +838,10 @@ export class Engine extends IEngine {
     if (this.client.session.keys.includes(topic)) {
       const clientRpcId = payloadId();
       const relayRpcId = getBigIntRpcId().toString() as any;
-      const { done, resolve, reject } = createDelayedPromise<void>();
+      const { done, resolve, reject } = createDelayedPromise<void>(
+        FIVE_MINUTES,
+        "Ping request expired without receiving any acknowledgement",
+      );
       this.events.once(engineEvent("session_ping", clientRpcId), ({ error }: any) => {
         if (error) reject(error);
         else resolve();
@@ -1466,15 +1514,17 @@ export class Engine extends IEngine {
     this.client.core.storage
       .removeItem(WALLETCONNECT_DEEPLINK_CHOICE)
       .catch((e) => this.client.logger.warn(e));
-    this.getPendingSessionRequests().forEach((r) => {
-      if (r.topic === topic) {
-        this.deletePendingSessionRequest(r.id, getSdkError("USER_DISCONNECTED"));
-      }
-    });
     // reset the queue state back to idle if a request for the deleted session is still in the queue
     if (topic === this.sessionRequestQueue.queue[0]?.topic) {
       this.sessionRequestQueue.state = ENGINE_QUEUE_STATES.idle;
     }
+
+    await Promise.all(
+      this.getPendingSessionRequests()
+        .filter((r) => r.topic === topic)
+        .map((r) => this.deletePendingSessionRequest(r.id, getSdkError("USER_DISCONNECTED"))),
+    );
+
     if (emitEvent) this.client.events.emit("session_delete", { id, topic });
   };
 
@@ -1879,10 +1929,14 @@ export class Engine extends IEngine {
         await this.onRelayEventResponse({ topic, payload, transportType });
         this.client.core.history.delete(topic, payload.id);
       } else {
+        this.client.logger.error(`onRelayMessage() -> unknown payload: ${JSON.stringify(payload)}`);
         await this.onRelayEventUnknownPayload({ topic, payload, transportType });
       }
       await this.client.core.relayer.messages.ack(topic, message);
     } catch (error) {
+      this.client.logger.error(
+        `onRelayMessage() -> failed to process an inbound message: ${message}`,
+      );
       this.client.logger.error(error);
     }
   }
@@ -2131,8 +2185,7 @@ export class Engine extends IEngine {
         sessionProperties,
         scopedProperties,
         sessionConfig,
-        authentication,
-        walletPayResult,
+        proposalRequestsResponses,
       } = payload.params;
       const pendingSession = [...this.pendingSessions.values()].find(
         (s) => s.sessionTopic === topic,
@@ -2166,8 +2219,8 @@ export class Engine extends IEngine {
         ...(scopedProperties && { scopedProperties }),
         ...(sessionConfig && { sessionConfig }),
         transportType: TRANSPORT_TYPES.relay,
-        authentication,
-        walletPayResult,
+        authentication: proposalRequestsResponses?.authentication,
+        walletPayResult: proposalRequestsResponses?.walletPay,
       };
 
       await this.client.session.set(session.topic, session);
@@ -2178,9 +2231,6 @@ export class Engine extends IEngine {
         metadata: session.peer.metadata,
       });
 
-      this.client.events.emit("session_connect", { session });
-      this.events.emit(engineEvent("session_connect", pendingSession.proposalId), { session });
-
       this.pendingSessions.delete(pendingSession.proposalId);
       this.deleteProposal(pendingSession.proposalId, false);
       this.cleanupDuplicatePairings(session);
@@ -2188,8 +2238,12 @@ export class Engine extends IEngine {
       await this.sendResult<"wc_sessionSettle">({
         id: payload.id,
         topic,
+        throwOnFailedPublish: true,
         result: true,
       });
+
+      this.client.events.emit("session_connect", { session });
+      this.events.emit(engineEvent("session_connect", pendingSession.proposalId), { session });
     } catch (err: any) {
       await this.sendError({
         id,
@@ -2360,21 +2414,9 @@ export class Engine extends IEngine {
   ) => {
     const { id } = payload;
     try {
-      this.isValidDisconnect({ topic, reason: payload.params });
-      await Promise.all([
-        new Promise((resolve) => {
-          // RPC request needs to happen before deletion as it utilizes session encryption
-          this.client.core.relayer.once(RELAYER_EVENTS.publish, async () => {
-            resolve(await this.deleteSession({ topic, id }));
-          });
-        }),
-        this.sendResult<"wc_sessionDelete">({
-          id,
-          topic,
-          result: true,
-        }),
-        this.cleanupPendingSentRequestsForTopic({ topic, error: getSdkError("USER_DISCONNECTED") }),
-      ]).catch((err) => this.client.logger.error(err));
+      await this.isValidDisconnect({ topic, reason: payload.params });
+      this.cleanupPendingSentRequestsForTopic({ topic, error: getSdkError("USER_DISCONNECTED") });
+      await this.deleteSession({ topic, id });
     } catch (err: any) {
       this.client.logger.error(err);
     }
@@ -2586,12 +2628,6 @@ export class Engine extends IEngine {
         (r) => r.topic === topic && r.request.method === "wc_sessionRequest",
       );
       forSession.forEach((r) => {
-        const id = r.request.id;
-        const target = engineEvent("session_request", id);
-        const listeners = this.events.listenerCount(target);
-        if (listeners === 0) {
-          throw new Error(`emitting ${target} without any listeners`);
-        }
         // notify .request() handler of the rejection
         this.events.emit(engineEvent("session_request", r.request.id), {
           error,
@@ -3316,6 +3352,26 @@ export class Engine extends IEngine {
     }
   };
 
+  private getTVFApproveParams = (session: SessionTypes.Struct) => {
+    try {
+      const approvedChains = getNamespacesChains(session.namespaces);
+      const approvedMethods = getNamespacesMethods(session.namespaces);
+      const approvedEvents = getNamespacesEvents(session.namespaces);
+      const sessionProperties = session.sessionProperties;
+      const scopedProperties = session.scopedProperties;
+      return {
+        approvedChains,
+        approvedMethods,
+        approvedEvents,
+        sessionProperties,
+        scopedProperties,
+      };
+    } catch (e) {
+      this.client.logger.warn(e, "Error getting TVF approve params");
+      return {};
+    }
+  };
+
   private getTVFParams = (
     id: number,
     params: JsonRpcTypes.RequestParams["wc_sessionRequest"],
@@ -3431,29 +3487,5 @@ export class Engine extends IEngine {
       this.client.logger.warn(e, "Error extracting tx hashes from result");
     }
     return [];
-  };
-
-  private prepareProposalRequestsResponses = (
-    proposalRequestsResponses: EngineTypes.ApproveParams["proposalRequestsResponses"] = [],
-  ) => {
-    const authentication: AuthTypes.Cacao[] = [];
-    const walletPayResult: EngineTypes.WalletPayResult[] = [];
-
-    proposalRequestsResponses.forEach((response) => {
-      try {
-        if (typeof response === "object" && (response as AuthTypes.Cacao)?.h && "h" in response) {
-          authentication.push(response);
-        } else if (
-          typeof response === "object" &&
-          (response as EngineTypes.WalletPayResult)?.txid &&
-          "txid" in response
-        ) {
-          walletPayResult.push(response);
-        }
-      } catch (e) {
-        this.client.logger.warn(e, "Error preparing proposal requests responses");
-      }
-    });
-    return { authentication, walletPayResult };
   };
 }