+++ packages/types/src/sign-client/engine.ts
@@ -105,6 +105,11 @@ export declare namespace EngineTypes {
     uri: string;
   }
 
+  type ProposalRequestsResponses = {
+    authentication?: AuthTypes.Cacao[];
+    walletPay?: WalletPayResult[];
+  };
+
   interface ApproveParams {
     id: number;
     namespaces: SessionTypes.Namespaces;
@@ -112,7 +117,7 @@ export declare namespace EngineTypes {
     scopedProperties?: ProposalTypes.ScopedProperties;
     sessionConfig?: SessionTypes.SessionConfig;
     relayProtocol?: string;
-    proposalRequestsResponses?: (AuthTypes.Cacao | WalletPayResult)[];
+    proposalRequestsResponses?: ProposalRequestsResponses;
   }
 
   interface RejectParams {